coresight: axim: Added driver for the AXI monitor
The AXI monitors is a bus monitor that exists in both AP-806 and
CP-110. it is capable of detecting specific bus transactions and
generate trace events or increment counters

This driver adds the basic support for the AXI monitor. it enables
using the AXI monitor for debug purposes. future patches will add
support for tracing using the axi monitor

Documentation/trace/axi-monitor.txt provides a basic explanation on
the monitor on how to use this driver

Documentation/ABI/testing/sysfs-bus-coresight-devices-axim gives
detailed descirption of all the sysfs file needed to operate the AXI
monitor

Change-Id: Ibad0250ee08c37538c4132d0cdad1dae87e03b78
Signed-off-by: Omri Itach <omrii@marvell.com>