net: mvpp2: use dynamic search for HASH lookup entries
Modify the way classifier HASH lookups are accessed.
Before this patch, the lookups were expected to be in
a predefined place in the flow sequence.
This patch allows to search for the HASH lookups, providing
more flexibility to re-arrange the HASH lookups.
This is useful i.e. when adding further lookups into a
sequence, usually the lookups are grouped per engine
(all C3 together, all C2 together).

Change-Id: Ibf106df8cbfce9f7748b70149b342b6c9a8dedab
Reviewed-on: http://vgitil04.il.marvell.com:8080/61211
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>