net: mvpp2: force TX FIFO transactions priority on the AXI bus
It may happen that when there are processes which load
the internal AXI bus with CPU-DRAM transactions, it is
possible to observe the TX FIFO underrun errors. It means,
that the PP2 port is unable to fetch the data from DRAM
on time during egress.

In order to avoid such issue, prioritize TX FIFO AXI bus
access using the PP2 NIC AXI bridge configuration.

Change-Id: I3d36776e911396e20254a319c3e8f87d76a23e61
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/1177
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>