dmatest: Enable DMA performance measurement using dmatest
This commit include all 4.4 commits regarding DMA performance measurement.
Reuse dmatest infrastructure to perform testing on DMA engine performance.
This adds several parameters to dmatest to control:
- Whether we are running in performance measurement mode or not.
- Number of outstanding requests we send to DMA engine before waiting for
  the first response.
- Specify if the test should be done using SW memory operations
  (e.g. memcpy) or the DMA offload engine.
- Whether the source buffers used for the memory operations are
  cache-resident or not.

This patch simply adds a new dmatest thread that is run instead of the
original dmatest thread in case we are running in performance measurement
mode.

Change-Id: I9a7152ee6b511083076382f9c2f1093e6be82681
Signed-off-by: Ben Peled <bpeled@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/5477
Tested-by: Stefan Chulski <Stefan.Chulski@cavium.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>